Add possibility to work on Pacific ocean.
Added a way to change the longitude coordinates from -180° to 180°, to 0° to 360°.
One need to use a lon_min value higher than a lon_max value.
For example ```lon_min = 150``` and ````lon_max = -150```

**[WARNING]** the use of the xarray function roll in the loop increses the runtime drastically.
I would suggest running for the whole grid then compining the netcdf files and change the coordinates.

Mostly  useful for the synthetic_CCF.ipynb

On branch dev_LT

Changelog: Added